Scaffold ShapeQueries
