+++ diff/src/queries/mod.rs
@@ -1,1 +1,2 @@
 pub mod endpoint;
+pub mod shape;
