+++ diff/src/queries/shape.rs
@@ -1,0 +1,42 @@
+use crate::projections::shape::ShapeProjection;
+use crate::shapes::traverser::{ShapeTrail, ShapeTrailPathComponent};
+use crate::state::shape::ShapeKind;
+
+pub struct ShapeQueries<'a> {
+  pub shape_projection: &'a ShapeProjection,
+}
+
+impl<'a> ShapeQueries<'a> {
+  pub fn new(shape_projection: &'a ShapeProjection) -> Self {
+    ShapeQueries { shape_projection }
+  }
+
+  pub fn list_trail_choices(
+    &self,
+    shape_trail: &ShapeTrail,
+    /*, parameter bindings */
+  ) -> Vec<ChoiceOutput> {
+    let graph = &self.shape_projection.graph;
+
+    vec![]
+  }
+}
+
+pub struct ChoiceOutput {
+  pub parent_trail: ShapeTrail,
+  pub additional_components: Vec<ShapeTrailPathComponent>,
+  // shape_id: ShapeId,
+  pub core_shape_kind: ShapeKind,
+}
+
+impl ChoiceOutput {
+  pub fn shape_trail(&self) -> ShapeTrail {
+    let mut path = self.parent_trail.path.clone();
+    let mut additional_components = self.additional_components.clone();
+    path.append(&mut additional_components);
+    ShapeTrail {
+      root_shape_id: self.parent_trail.root_shape_id.clone(),
+      path,
+    }
+  }
+}
